Exceptions and Loggers (XML settings) Added, converted to Maven Project with POM
