+++ logs/app.log
@@ -1,0 +1,1 @@
+2023-26-02 03:26:24.914 [main] INFO  com.solvd.laba.InstantiateObjects - hello
