--- src/NonMedicalWorkers/CleaningAndSanitization.java
@@ -1,17 +1,0 @@
-package NonMedicalWorkers;
-
-import MedicalWorkers.GeneralSurgeon;
-
-public interface CleaningAndSanitization {
-	
-	
-	public final String department = "Cleaning and Sanitization";
-	
-	public void SanitizeSpace();
-	
-	public void identifySelf();
-
-	public int getAge(); 
-	
-	public String getName();
-}
