--- src/Patients/Patient.java
@@ -1,10 +1,0 @@
-package Patients;
-
-public interface Patient {
-
-	public void checkIn();
-	
-	public void checkOut();
-
-	public void fillOutPatientForm(String time);
-}
