--- src/Tools/Tools.java
@@ -1,10 +1,0 @@
-package Tools;
-
-public interface Tools {
-
-	public static final String ownerPermisson = "Doctor";
-	
-	public void cleanTool();
-	
-	public void equipTool();
-}
