+++ src/main/java/com/solvd/laba/Exceptions/PatientNotFoundException.java
@@ -1,0 +1,18 @@
+package com.solvd.laba.Exceptions;
+
+public class PatientNotFoundException extends Exception {
+
+	private static final long serialVersionUID = 1L;
+	
+	public PatientNotFoundException() {
+	}
+	
+	public PatientNotFoundException(String message) {
+		super(message);
+	}
+	
+	public PatientNotFoundException(String message, Throwable cause) {
+		super(message, cause);
+		
+}
+}
