+++ src/main/java/com/solvd/laba/Exceptions/UnregisteredSurgeryException.java
@@ -1,0 +1,18 @@
+package com.solvd.laba.Exceptions;
+
+public class UnregisteredSurgeryException extends Exception{
+
+	private static final long serialVersionUID = 1L;
+
+	public UnregisteredSurgeryException() {
+	}
+	
+	public UnregisteredSurgeryException(String message) {
+		super(message);
+	}
+	
+	public UnregisteredSurgeryException(String message, Throwable cause) {
+		super(message, cause);
+		
+}
+}
