+++ src/main/java/com/solvd/laba/NonMedicalWorkers/CleaningAndSanitization.java
@@ -1,0 +1,15 @@
+package com.solvd.laba.NonMedicalWorkers;
+
+public interface CleaningAndSanitization {
+	
+	
+	public final String department = "Cleaning and Sanitization";
+	
+	public void SanitizeSpace();
+	
+	public void identifySelf();
+
+	public int getAge(); 
+	
+	public String getName();
+}
