+++ src/main/java/com/solvd/laba/Patients/Patient.java
@@ -1,0 +1,12 @@
+package com.solvd.laba.Patients;
+
+import java.io.IOException;
+
+public interface Patient {
+
+	public void checkIn();
+	
+	public void checkOut();
+
+	public void fillOutPatientForm(String time) throws IOException;
+}
